Fix temp file paths used in tests (#514)
## Scope

Running tests with nose would use the `tests` directory as CWD, so test file paths were all relative to the `tests` directory.  In contrast, pytest runs from the `pysam` base directory as CWD, so all test file file paths must be updated.  Prior changes were sufficient to ensure that existing test files were found in the correct locations, but many temporary files were created in CWD, rather than the `tests` directory.  

## Limitations

- There are still a few temp files not created in `tests`, including the 1kGenomes index file.
- In future, all temporary files should probably be written to a dedicated directory that can be removed after the test suite is run.

+++ tests/tabix_test.py
@@ -138,13 +138,12 @@ class TestCompression(unittest.TestCase):
         checkBinaryEqual(self.tmpfilename + ".gz.tbi", self.filename_idx)
 
     def tearDown(self):
-
-        try:
+        if os.path.exists(self.tmpfilename):
             os.unlink(self.tmpfilename)
+        if os.path.exists(self.tmpfilename + ".gz"):
             os.unlink(self.tmpfilename + ".gz")
+        if os.path.exists(self.tmpfilename + ".gz.tbi"):
             os.unlink(self.tmpfilename + ".gz.tbi")
-        except OSError:
-            pass
 
 
 class TestCompressionSam(TestCompression):
